feat: add UseTags custom hook
