+++ src/pages/MainPage.tsx
@@ -1,4 +1,3 @@
-import { collection, getDocs } from '@firebase/firestore';
 import React, { memo, useEffect, useState } from 'react';
 
 import Form from 'components/Form';
@@ -9,11 +8,8 @@ import MainPageImage from 'components/SvgElements/MainPageImage';
 import TagsInput from 'components/TagsInput';
 
 import { sortByDate } from 'utils/functions';
-import { Tag } from 'utils/interfaces';
 
 import useNotes from 'hooks/useNotes';
-
-import dataBase from '../firebase';
 
 function MainPage() {
   const [isLoading, setLoading] = useState(false);
@@ -40,20 +36,6 @@ function MainPage() {
     })();
   }, []);
 
-  const [tags, setTags] = useState<Tag[]>([]);
-
-  const tagsCollection = collection(dataBase, 'tags');
-  useEffect(() => {
-    (async () => {
-      const data = await getDocs(tagsCollection);
-      const tagsFromDataBase: Tag[] = data.docs.map((item) => ({
-        ...item.data(),
-        id: item.id,
-      }));
-      setTags(tagsFromDataBase);
-    })();
-  }, []);
-
   return (
     <main>
       <section className="notes-creation">
@@ -71,7 +53,7 @@ function MainPage() {
         </div>
       </section>
       <section className="tags-container">
-        <TagsInput tags={tags} setTags={setTags} />
+        <TagsInput />
       </section>
       <section
         className={notes.length ? 'notes-container' : 'notes-container_empty'}
